remove dependence on global post
for adaptive mode.  In the block context, the root of the navigation won't always be the global post object.  This is also true for API requests.  For performance reasons, this change will still use the global post if it matches the request parameters.  Otherwise, this uses a new parameter added to the 'adaptive_pages_filter' function to determine the root post.